veerrzastore/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
netaddr==1.3.0
psycopg2==2.9.10
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
wrapt==1.17.2
yookassa==3.5.0
